requirements.txt
# requirements.txt 
PyAutoGUI==0.9.52
Pillow==8.4.0

pandas==2.8.2
opencv-python-headless==4.6.0.66
streamlit==1.12.2

